+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 This file keeps track of changes between releases for the ruolo project
 which adheres to [semantic versioning](https://semver.org).
 
+## unreleased
+
+* Add a `role?` method to the user mixin to check for given roles instead
+  of permissions.
+
 ## v0.2.0 2019-09-11
 
 Initial cleanup of original code: add tests and documentation.